+++ aitools/proofs/knowledge_bases/zodb.py
@@ -7,7 +7,8 @@ from persistent.mapping import PersistentMapping
 
 from aitools.logic import Expression, Substitution, Variable
 from aitools.logic.utils import normalize_variables, VariableSource
-from aitools.proofs.index import AbstruseIndex, TrieIndex, make_key
+from aitools.proofs.index import make_key
+from aitools.utils.abstruse_index import TrieIndex, AbstruseIndex
 from aitools.proofs.knowledge_bases.knowledge_base import KnowledgeBase
 from aitools.proofs.listeners import Listener
 from aitools.proofs.proof import Prover
